requirements.txt
camelot_py==0.11.0
pandas==1.5.0
PySimpleGUI==4.60.5
